+++ test/swift/apps/experimental/ViewController.swift
@@ -17,8 +17,6 @@ final class ViewController: UITableViewController {
   override func viewDidLoad() {
     super.viewDidLoad()
 
-    let userDefaults = UserDefaults.standard
-
     let engine = EngineBuilder()
       .addLogLevel(.debug)
       .addPlatformFilter(DemoFilter.init)
@@ -37,8 +35,7 @@ final class ViewController: UITableViewController {
       )
       .setOnEngineRunning { NSLog("Envoy async internal setup completed") }
       .addStringAccessor(name: "demo-accessor", accessor: { return "PlatformString" })
-      .addKeyValueStore(name: "demo-kv-store",
-                        keyValueStore: UserDefaultsStore(userDefaults: userDefaults))
+      .addKeyValueStore(name: "demo-kv-store", keyValueStore: UserDefaults.standard)
       .setEventTracker { NSLog("Envoy event emitted: \($0)") }
       .forceIPv6(true)
       .build()